+++ utils/data.py
@@ -52,7 +52,7 @@ class Message(BaseModel):
     message_id: int | str = None    # str if id is temporary
     m_type: MessageType = MessageType.TEXT
     t_type: TargetType = TargetType.OTHER
-    time: float = None  # write by backend
+    time: int = None  # write by backend
     content: str | list | int | GroupData | UserData  # 如果是消息，content 是 str，如果是函数，content 是 list,如果是群加人，这个放群id
     sender: int = None  # 如果是消息，sender 是发送者的 id，如果是函数，sender 是函数的发起者的 id。如果是群加人，这个放拉人的人
     receiver: int | None = None  # 如果是消息，receiver 是接收者的 id，如果是函数，receiver 是函数的接收者的 id。如果是群加人，这个放被拉的人
